CTCP-4862: Organising answers helper packages.
