+++ app/utils/answersHelpers/consignment/transportEquipment/SealAnswersHelper.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package utils.answersHelpers
+package utils.answersHelpers.consignment.transportEquipment
 
 import models.{Index, UserAnswers}
 import pages.transportEquipment.index.seals.SealIdentificationNumberPage
@@ -22,6 +22,7 @@ import play.api.i18n.Messages
 import play.api.mvc.Call
 import uk.gov.hmrc.govukfrontend.views.viewmodels.summarylist.SummaryListRow
 import uk.gov.hmrc.http.HttpVerbs.GET
+import utils.answersHelpers.AnswersHelper
 
 class SealAnswersHelper(
   userAnswers: UserAnswers,
